requirements.txt
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==5.5
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
idna==3.6
lyricsgenius==3.0.1
mutagen==1.47.0
packaging==24.0
pyaes==1.6.1
pyasn1==0.6.0
pycryptodomex==3.20.0
python-decouple==3.8
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
Telethon==1.34.0
toml==0.10.2
urllib3==2.2.1
websockets==12.0
youtube-search==2.1.2
yt-dlp>=2023.12.30
sqlalchemy
psycopg2-binary==2.9.9
psycopg2
pytz==2023.3
telethon
spotipy
youtube-search
yt-dlp
requests
